feat: changed file structure to separate graph and list and improved styles usage
