+++ screens/Contacts/ContactGraph/ContactGraph.tsx
@@ -1,0 +1,12 @@
+import { View, Text } from "react-native"
+import { styles } from "./styles"
+
+const ContactGraph = () => {
+  return (
+    <View style={styles.container}>
+      <Text>profile interests</Text>
+    </View>
+  )
+}
+
+export default ContactGraph
