+++ screens/Contacts/ContactGraph/styles.ts
@@ -1,0 +1,11 @@
+import { StyleSheet } from "react-native" 
+
+export const styles = StyleSheet.create({
+    container: {
+        alignItems: "center",
+        flex: 1,
+        justifyContent: "center",
+    },
+
+})
+
